fix(android): merging example for ios and android
